Add some message-dedupe to the send queues
Maintain a local set of all currently queued message IDs, and don't requeue messages with
an ID that are currently in the queue.

As this set is local per-sender/slave, it will only prevent messages that are taking a *long*
time to send from being sent twice if the same ID hits the same sender more than once. Eg if
we have two slaves, at most a message will be sent twice (one per slave) but subsequent messages
will not be sent as both of those slaves will have the message ID in their set.

We could use zookeeper for a dedupe functionality to be maintained across the masters and slaves,
but that would entail significant additional usage on our ZK cluster which might not be
ideal.